+++ dali-toolkit/internal/visuals/wireframe/wireframe-visual.cpp
@@ -42,7 +42,7 @@ const char * const INDEX_NAME("indices");
 
 const char* VERTEX_SHADER = DALI_COMPOSE_SHADER(
 attribute mediump vec2  aPosition;\n
-uniform   mediump mat4  uMvpMatrix;\n
+uniform   highp   mat4  uMvpMatrix;\n
 uniform   mediump vec3  uSize;\n
 \n
